+++ src/errors/index.ts
@@ -4,6 +4,7 @@ export * from "./ArgumentValidationError";
 export * from "./CannotDetermineTypeError";
 export * from "./ForbiddenError";
 export * from "./GeneratingSchemaError";
+export * from "./ConflictingDefaultValuesError";
 export * from "./MissingSubscriptionTopicsError";
 export * from "./NoExplicitTypeError";
 export * from "./ReflectMetadataMissingError";